+++ nionswift_plugin/DM_IO/__init__.py
@@ -11,8 +11,7 @@ import typing
 from nion.data import DataAndMetadata
 
 # local libraries
-from . import dm3_image_utils
-
+from nion.io.DM_IO import dm3_image_utils
 
 _ = gettext.gettext
