+++ pyproject.toml
@@ -32,16 +32,17 @@ dependencies = [
 include = [
     "nionswift_plugin",
     "nionswift_plugin.DM_IO",
-    "nionswift_plugin.DM_IO.test",
+    "nion.io.DM_IO",
+    "nion.io.DM_IO.test",
     "nionswift_plugin.TIFF_IO",
     "nionswift_plugin.TIFF_IO.test"
 ]
 
 [tool.setuptools.package-data]
-"nionswift_plugin.DM_IO.test" = ["resources/*"]
+"nion.io.DM_IO.test" = ["resources/*"]
 
 [tool.pytest.ini_options]
 testpaths = [
-    "nionswift_plugin/DM_IO",
+    "nion/io/DM_IO",
     "nionswift_plugin/TIFF_IO"
 ]
